All free from warngings.

+++ cppcache/integration-test/testOverflowPutGetSqLite.cpp
@@ -88,7 +88,7 @@ void checkOverflowTokenValues(std::shared_ptr<Region>& regionPtr,
   std::shared_ptr<CacheableKey> keyPtr;
   std::shared_ptr<Cacheable> valuePtr;
   int count = 0;
-  int nonoverflowCount = 0;
+  size_t nonoverflowCount = 0;
   for (uint32_t i = 0; i < num; i++) {
     keyPtr = v.at(i);
     auto rPtr = regionPtr->getEntry(keyPtr);
