+++ cppcache/integration-test/testThinClientPoolExecuteHAFunctionPrSHOP.cpp
@@ -228,10 +228,10 @@ DUNIT_TASK_DEFINITION(CLIENT1, Client1OpTest)
         LOG(buf);
         ASSERT(resultList->size() == 17,
                "get executeFunctionResult count is not 17");
-        for (int32_t i = 0; i < resultList->size(); i++) {
-          sprintf(buf, "result[%d] is null\n", i);
+        for (size_t i = 0; i < resultList->size(); i++) {
+          sprintf(buf, "result[%zd] is null\n", i);
           ASSERT(resultList->operator[](i) != nullptr, buf);
-          sprintf(buf, "get result[%d]=%s", i,
+          sprintf(buf, "get result[%zd]=%s", i,
                   std::dynamic_pointer_cast<CacheableString>(
                       resultList->operator[](i))
                       ->value()
@@ -309,10 +309,10 @@ DUNIT_TASK_DEFINITION(CLIENT1, Client1OnServerHATest)
         LOG(buf);
         ASSERT(resultList->size() == 17,
                "get executeFunctionResult count is not 17");
-        for (int32_t i = 0; i < resultList->size(); i++) {
-          sprintf(buf, "result[%d] is null\n", i);
+        for (size_t i = 0; i < resultList->size(); i++) {
+          sprintf(buf, "result[%zd] is null\n", i);
           ASSERT(resultList->operator[](i) != nullptr, buf);
-          sprintf(buf, "get result[%d]=%s", i,
+          sprintf(buf, "get result[%zd]=%s", i,
                   std::dynamic_pointer_cast<CacheableString>(
                       resultList->operator[](i))
                       ->value()
@@ -349,10 +349,10 @@ DUNIT_TASK_DEFINITION(CLIENT1, Client1OnServerHATest)
         LOG(buf);
         ASSERT(resultList->size() == 17,
                "get executeFunctionResult1 count is not 17");
-        for (int32_t i = 0; i < resultList->size(); i++) {
-          sprintf(buf, "result[%d] is null\n", i);
+        for (size_t i = 0; i < resultList->size(); i++) {
+          sprintf(buf, "result[%zd] is null\n", i);
           ASSERT(resultList->operator[](i) != nullptr, buf);
-          sprintf(buf, "get result[%d]=%s", i,
+          sprintf(buf, "get result[%zd]=%s", i,
                   std::dynamic_pointer_cast<CacheableString>(
                       resultList->operator[](i))
                       ->value()
